removes network code from main activity
